// backend/src/application/offers/types/OfferReturnType.d.ts
export type OfferReturnType = {
    uuid: string,
    ownerUuid: string,
    title: string,
    body: string,
    contactDetails: { 
        phoneNumber: string,
        address: string,
        email: string,
        website: string,
        github: string,
        linkedin: string
    },
    company: {
        name: string,
        size: string,
        website: string,
        industry: string
    },
    location: {
        city: string,
        country: string,
        modality: string
    },
    salary: {
        min: number,
        max: number,
        currency: string,
        period: string,
        equity: boolean
    },
    requirements: Set<string>
}